requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.1.0=pypi_0
anyio=3.6.1=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pypi_0
babel=2.10.3=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
bleach=5.0.0=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
ca-certificates=2021.7.5=h06a4308_1
cachetools=5.2.0=pypi_0
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chardet=4.0.0=py38h06a4308_1003
conda=4.10.3=py38h06a4308_0
conda-package-handling=1.7.3=py38h27cfd23_1
cryptography=3.4.7=py38hd23ed53_0
cycler=0.11.0=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.5.1=pypi_0
dominate=2.6.0=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fastjsonschema=2.15.3=pypi_0
fonttools=4.33.3=pypi_0
google-auth=2.8.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.46.3=pypi_0
idna=2.10=pyhd3eb1b0_0
imageio=2.19.3=pypi_0
importlib-metadata=4.11.4=pypi_0
importlib-resources=5.8.0=pypi_0
ipykernel=6.15.0=pypi_0
ipython=8.4.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
jittor=1.3.4.16=pypi_0
json5=0.9.8=pypi_0
jsonschema=4.6.0=pypi_0
jupyter-client=7.3.4=pypi_0
jupyter-core=4.10.0=pypi_0
jupyter-server=1.17.1=pypi_0
jupyterlab=3.4.3=pypi_0
jupyterlab-language-pack-zh-cn=3.4.post1=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.14.0=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.4.3=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
markdown=3.3.7=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
nbclassic=0.3.7=pypi_0
nbclient=0.6.4=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.5=pypi_0
networkx=2.8.4=pypi_0
notebook=6.4.12=pypi_0
notebook-shim=0.1.0=pypi_0
numpy=1.22.4=pypi_0
oauthlib=3.2.0=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1k=h27cfd23_0
packaging=21.3=pypi_0
pandas=1.4.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.1.1=pypi_0
pip=21.1.3=py38h06a4308_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycosat=0.6.3=py38h7b6447c_1
pycparser=2.20=py_2
pygments=2.12.0=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pypi_0
pytz=2022.1=pypi_0
pywavelets=1.3.0=pypi_0
pyzmq=23.2.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
ruamel_yaml=0.15.100=py38h27cfd23_0
scikit-image=0.19.3=pypi_0
scipy=1.8.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=52.0.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.36.0=hc218d9a_0
stack-data=0.3.0=pypi_0
supervisor=4.2.4=pypi_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
terminado=0.15.0=pypi_0
tifffile=2022.5.4=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.10=hbc83047_0
torch=1.11.0+cu113=pypi_0
torch-tb-profiler=0.4.0=pypi_0
torchvision=0.12.0+cu113=pypi_0
tornado=6.1=pypi_0
tqdm=4.61.2=pyhd3eb1b0_1
traitlets=5.3.0=pypi_0
typing-extensions=4.2.0=pypi_0
urllib3=1.26.6=pyhd3eb1b0_1
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.3=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.8.0=pypi_0
zlib=1.2.11=h7b6447c_3
